CH-59 Rename __APP_NAME__ directory to ch_django and api directory to __APP_NAME__
