+++ application-templates/django-app/backend/__APP_NAME__/apps.py
