+++ application-templates/django-app/backend/ch_django/models.py
