+++ application-templates/django-app/backend/ch_django/settings.py
