+++ application-templates/django-app/backend/ch_django/static/www/index.html
