+++ application-templates/django-app/backend/ch_django/templates/__APP_NAME__/swagger-ui.html
