+++ application-templates/django-app/backend/ch_django/urls.py
